// src/modules/planners/models/planners-messages.enum.ts
export enum PlannersSocketMessages {
  PRESENCE_SNAPSHOT = 'PRESENCE_SNAPSHOT',
  USER_JOINED = 'USER_JOINED',
  USER_LEFT = 'USER_LEFT',
  INSTANCE_ADDED = 'INSTANCE_ADDED',
  INSTANCE_UPDATED = 'INSTANCE_UPDATED',
  INSTANCE_REMOVED = 'INSTANCE_REMOVED',
  LEGEND_ADDED = 'LEGEND_ADDED',
  LEGEND_UPDATED = 'LEGEND_UPDATED',
  LEGEND_REMOVED = 'LEGEND_REMOVED',
}

export enum PlannersClientMessages {
  JOIN_PLANNER = 'JOIN_PLANNER',
  ADD_INSTANCE = 'ADD_INSTANCE',
  UPDATE_INSTANCE = 'UPDATE_INSTANCE',
  REMOVE_INSTANCE = 'REMOVE_INSTANCE',
  ADD_LEGEND = 'ADD_LEGEND',
  UPDATE_LEGEND = 'UPDATE_LEGEND',
  REMOVE_LEGEND = 'REMOVE_LEGEND',
}